+++ Source/Scene/LabelCollection.js
@@ -1,6 +1,8 @@
 /*global define*/
 define([
+        '../Core/BoundingRectangle',
         '../Core/Cartesian2',
+        '../Core/Color',
         '../Core/defaultValue',
         '../Core/defined',
         '../Core/defineProperties',
@@ -16,7 +18,9 @@ define([
         './TextureAtlas',
         './VerticalOrigin'
     ], function(
+        BoundingRectangle,
         Cartesian2,
+        Color,
         defaultValue,
         defined,
         defineProperties,
@@ -53,6 +57,24 @@ define([
         this.dimensions = dimensions;
     }
 
+    var whitePixelCanvasId = 'ID_WHITE_PIXEL';
+    var whitePixelSize = new Cartesian2(4, 4);
+    var whitePixelBoundingRegion = new BoundingRectangle(1, 1, 1, 1);
+
+    function addWhitePixelCanvas(textureAtlas, labelCollection) {
+        var canvas = document.createElement('canvas');
+        canvas.width = whitePixelSize.x;
+        canvas.height = whitePixelSize.y;
+
+        var context2D = canvas.getContext('2d');
+        context2D.fillStyle = '#fff';
+        context2D.fillRect(0, 0, canvas.width, canvas.height);
+
+        textureAtlas.addImage(whitePixelCanvasId, canvas).then(function(index) {
+            labelCollection._whitePixelIndex = index;
+        });
+    }
+
     // reusable object for calling writeTextToCanvas
     var writeTextToCanvasParameters = {};
     function createGlyphCanvas(character, font, fillColor, outlineColor, outlineWidth, style, verticalOrigin) {
@@ -61,13 +83,13 @@ define([
         writeTextToCanvasParameters.strokeColor = outlineColor;
         writeTextToCanvasParameters.strokeWidth = outlineWidth;
 
-        if (verticalOrigin === VerticalOrigin.BOTTOM) {
-            writeTextToCanvasParameters.textBaseline = 'bottom';
+        if (verticalOrigin === VerticalOrigin.CENTER) {
+            writeTextToCanvasParameters.textBaseline = 'middle';
         } else if (verticalOrigin === VerticalOrigin.TOP) {
             writeTextToCanvasParameters.textBaseline = 'top';
         } else {
-            // VerticalOrigin.CENTER
-            writeTextToCanvasParameters.textBaseline = 'middle';
+            // VerticalOrigin.BOTTOM and VerticalOrigin.BASELINE
+            writeTextToCanvasParameters.textBaseline = 'bottom';
         }
 
         writeTextToCanvasParameters.fill = style === LabelStyle.FILL || style === LabelStyle.FILL_AND_OUTLINE;
@@ -114,6 +136,40 @@ define([
 
         // presize glyphs to match the new text length
         glyphs.length = textLength;
+
+        var showBackground = label._showBackground && (glyphs.length > 0);
+        var backgroundBillboard = label._backgroundBillboard;
+        var backgroundBillboardCollection = labelCollection._backgroundBillboardCollection;
+        if (!showBackground) {
+            if (defined(backgroundBillboard)) {
+                backgroundBillboardCollection.remove(backgroundBillboard);
+                label._backgroundBillboard = backgroundBillboard = undefined;
+            }
+        } else {
+            if (!defined(backgroundBillboard)) {
+                backgroundBillboard = backgroundBillboardCollection.add({
+                    collection : labelCollection,
+                    image : whitePixelCanvasId,
+                    imageSubRegion : whitePixelBoundingRegion
+                });
+                label._backgroundBillboard = backgroundBillboard;
+            }
+
+            backgroundBillboard.color = label._backgroundColor;
+            backgroundBillboard.show = label._show;
+            backgroundBillboard.position = label._position;
+            backgroundBillboard.eyeOffset = label._eyeOffset;
+            backgroundBillboard.pixelOffset = label._pixelOffset;
+            backgroundBillboard.horizontalOrigin = HorizontalOrigin.LEFT;
+            backgroundBillboard.verticalOrigin = label._verticalOrigin;
+            backgroundBillboard.heightReference = label._heightReference;
+            backgroundBillboard.scale = label._scale;
+            backgroundBillboard.pickPrimitive = label;
+            backgroundBillboard.id = label._id;
+            backgroundBillboard.translucencyByDistance = label._translucencyByDistance;
+            backgroundBillboard.pixelOffsetScaleByDistance = label._pixelOffsetScaleByDistance;
+            backgroundBillboard.distanceDisplayCondition = label._distanceDisplayCondition;
+        }
 
         var glyphTextureCache = labelCollection._glyphTextureCache;
 
@@ -179,9 +235,10 @@ define([
             // if we have a texture, configure the existing billboard, or obtain one
             if (glyphTextureInfo.index !== -1) {
                 var billboard = glyph.billboard;
+                var spareBillboards = labelCollection._spareBillboards;
                 if (!defined(billboard)) {
-                    if (labelCollection._spareBillboards.length > 0) {
-                        billboard = labelCollection._spareBillboards.pop();
+                    if (spareBillboards.length > 0) {
+                        billboard = spareBillboards.pop();
                     } else {
                         billboard = labelCollection._billboardCollection.add({
                             collection : labelCollection
@@ -212,46 +269,48 @@ define([
         label._repositionAllGlyphs = true;
     }
 
-    // reusable Cartesian2 instance
+    // reusable Cartesian2 instances
     var glyphPixelOffset = new Cartesian2();
+    var scratchBackgroundPadding = new Cartesian2();
 
     function repositionAllGlyphs(label, resolutionScale) {
-        
         var glyphs = label._glyphs;
         var text = label._text;
-
         var glyph;
         var dimensions;
-
         var lineWidth = 0;
         var maxLineWidth = 0;
-
-        var maxGlyphHeight = 0;
-        var maxGlyphDescent = 0;
-
+        var maxGlyphDescent = Number.NEGATIVE_INFINITY;
+        var maxGlyphY = 0;
         var numberOfLines = 1;
-
         var glyphIndex = 0;
         var glyphLength = glyphs.length;
-        for (glyphIndex = 0; glyphIndex < glyphLength; ++glyphIndex) {          
+
+        var backgroundBillboard = label._backgroundBillboard;
+        var backgroundPadding = scratchBackgroundPadding;
+        Cartesian2.clone(
+            (defined(backgroundBillboard) ? label._backgroundPadding : Cartesian2.ZERO),
+            backgroundPadding);
+
+        for (glyphIndex = 0; glyphIndex < glyphLength; ++glyphIndex) {
             if (text.charAt(glyphIndex) === '\n') {
-                numberOfLines += 1;
+                ++numberOfLines;
                 lineWidth = 0;
-            }
-            else {
+            } else {
                 glyph = glyphs[glyphIndex];
                 dimensions = glyph.dimensions;
+                maxGlyphY = Math.max(maxGlyphY, dimensions.height - dimensions.descent);
+                maxGlyphDescent = Math.max(maxGlyphDescent, dimensions.descent);
 
+                //Computing the line width must also account for the kerning that occurs between letters.
                 lineWidth += dimensions.width - dimensions.bounds.minx;
                 if (glyphIndex < glyphLength - 1) {
                     lineWidth += glyphs[glyphIndex + 1].dimensions.bounds.minx;
                 }
-
                 maxLineWidth = Math.max(maxLineWidth, lineWidth);
-                maxGlyphHeight = Math.max(maxGlyphHeight, dimensions.height);
-                maxGlyphDescent = Math.max(maxGlyphDescent, dimensions.descent);
             }
         }
+        var maxLineHeight = maxGlyphY + maxGlyphDescent;
 
         var scale = label._scale;
         var horizontalOrigin = label._horizontalOrigin;
@@ -259,15 +318,17 @@ define([
         if (horizontalOrigin === HorizontalOrigin.CENTER) {
             widthOffset -= maxLineWidth / 2 * scale;
         } else if (horizontalOrigin === HorizontalOrigin.RIGHT) {
-            widthOffset -= maxLineWidth * scale;
+            widthOffset -= (maxLineWidth + backgroundPadding.x) * scale;
+        } else {
+            widthOffset += backgroundPadding.x * scale;
         }
 
-
-        var maxLineHeight = maxGlyphHeight + maxGlyphDescent;
         var LEADING = 0.2*maxLineHeight; // Traditionally, leading is %20 of the font size.
         var totalTextHeight = (maxLineHeight * numberOfLines) + (LEADING * (numberOfLines-1));
 
         var topOffset = 0;
+        var verticalOrigin = label._verticalOrigin;
+        /*
         var heightReference = label._heightReference;
         var verticalOrigin = (heightReference === HeightReference.NONE) ? label._verticalOrigin : VerticalOrigin.BOTTOM;
         if (verticalOrigin === VerticalOrigin.CENTER) {
@@ -276,6 +337,7 @@ define([
             // Subtract maxGlyphHeight for backwards compatibility
             topOffset += (totalTextHeight * scale) - maxGlyphHeight;
         }
+        */
 
         glyphPixelOffset.x = widthOffset * resolutionScale;
         glyphPixelOffset.y = 0;
@@ -283,21 +345,24 @@ define([
 
         var glyphNewlineOffset = 0;
         for (glyphIndex = 0; glyphIndex < glyphLength; ++glyphIndex) {
-
             if (text.charAt(glyphIndex) === '\n') {
                 glyphNewlineOffset += (maxLineHeight + LEADING) * scale;
                 glyphPixelOffset.x = widthOffset;
                 continue;
             }
-            
+
             glyph = glyphs[glyphIndex];
             dimensions = glyph.dimensions;
-            if (verticalOrigin === VerticalOrigin.BOTTOM) {
+
+            if (verticalOrigin === VerticalOrigin.BASELINE) {
                 glyphPixelOffset.y = topOffset - dimensions.descent * scale;
             } else if (verticalOrigin === VerticalOrigin.TOP) {
-                glyphPixelOffset.y = topOffset - (maxGlyphHeight - dimensions.height) * scale - dimensions.descent * scale;
+                glyphPixelOffset.y = topOffset - (maxGlyphY - dimensions.height + dimensions.descent + backgroundPadding.y) * scale;
             } else if (verticalOrigin === VerticalOrigin.CENTER) {
-                glyphPixelOffset.y = topOffset - (maxGlyphHeight - dimensions.height) / 2 * scale - dimensions.descent * scale;
+                glyphPixelOffset.y = topOffset - (maxGlyphY - dimensions.height) / 2 * scale - dimensions.descent * scale;
+            } else {
+                // VerticalOrigin.BOTTOM
+                glyphPixelOffset.y = topOffset + (maxGlyphDescent - dimensions.descent + backgroundPadding.y) * scale;
             }
 
             glyphPixelOffset.y -= glyphNewlineOffset;
@@ -315,12 +380,34 @@ define([
                 glyphPixelOffset.x += ((dimensions.width - dimensions.bounds.minx) + nextGlyph.dimensions.bounds.minx) * scale * resolutionScale;
             }
         }
+
+        if (defined(backgroundBillboard) && (glyphLength > 0)) {
+            glyphPixelOffset.x = (widthOffset - backgroundPadding.x * scale) * resolutionScale;
+            if (verticalOrigin === VerticalOrigin.BASELINE) {
+                glyphPixelOffset.y = -backgroundPadding.y * scale - maxGlyphDescent * scale;
+            } else if (verticalOrigin === VerticalOrigin.TOP) {
+                glyphPixelOffset.y = -(maxGlyphY - maxLineHeight) * scale - maxGlyphDescent * scale;
+            } else if (verticalOrigin === VerticalOrigin.CENTER) {
+                glyphPixelOffset.y = -(maxGlyphY - maxLineHeight) / 2 * scale - maxGlyphDescent * scale;
+            } else {
+                // VerticalOrigin.BOTTOM
+                glyphPixelOffset.y = 0;
+            }
+            glyphPixelOffset.y *= resolutionScale;
+            backgroundBillboard.width = maxLineWidth + (backgroundPadding.x * 2);
+            backgroundBillboard.height = maxLineHeight + (backgroundPadding.y * 2);
+            backgroundBillboard._setTranslate(glyphPixelOffset);
+        }
     }
 
     function destroyLabel(labelCollection, label) {
         var glyphs = label._glyphs;
         for ( var i = 0, len = glyphs.length; i < len; ++i) {
             unbindGlyph(labelCollection, glyphs[i]);
+        }
+        if (defined(label._backgroundBillboard)) {
+            labelCollection._backgroundBillboardCollection.remove(label._backgroundBillboard);
+            label._backgroundBillboard = undefined;
         }
         label._labelCollection = undefined;
 
@@ -381,6 +468,13 @@ define([
         this._scene = options.scene;
 
         this._textureAtlas = undefined;
+        this._backgroundTextureAtlas = undefined;
+        this._whitePixelIndex = undefined;
+
+        this._backgroundBillboardCollection = new BillboardCollection({
+            scene : this._scene
+        });
+        this._backgroundBillboardCollection.destroyTextureAtlas = false;
 
         this._billboardCollection = new BillboardCollection({
             scene : this._scene
@@ -477,12 +571,20 @@ define([
      *   font : '30px sans-serif',
      *   fillColor : Cesium.Color.WHITE,
      *   outlineColor : Cesium.Color.BLACK,
+     *   outlineWidth : 1.0,
+     *   showBackground : false,
+     *   backgroundColor : new Cesium.Color(0.165, 0.165, 0.165, 0.8),
+     *   backgroundPadding : new Cesium.Cartesian2(7, 5),
      *   style : Cesium.LabelStyle.FILL,
      *   pixelOffset : Cesium.Cartesian2.ZERO,
      *   eyeOffset : Cesium.Cartesian3.ZERO,
      *   horizontalOrigin : Cesium.HorizontalOrigin.LEFT,
-     *   verticalOrigin : Cesium.VerticalOrigin.BOTTOM,
-     *   scale : 1.0
+     *   verticalOrigin : Cesium.VerticalOrigin.BASELINE,
+     *   scale : 1.0,
+     *   translucencyByDistance : undefined,
+     *   pixelOffsetScaleByDistance : undefined,
+     *   heightReference : HeightReference.NONE,
+     *   distanceDisplayCondition : undefined
      * });
      *
      * @example
@@ -623,9 +725,12 @@ define([
      */
     LabelCollection.prototype.update = function(frameState) {
         var billboardCollection = this._billboardCollection;
+        var backgroundBillboardCollection = this._backgroundBillboardCollection;
 
         billboardCollection.modelMatrix = this.modelMatrix;
         billboardCollection.debugShowBoundingVolume = this.debugShowBoundingVolume;
+        backgroundBillboardCollection.modelMatrix = this.modelMatrix;
+        backgroundBillboardCollection.debugShowBoundingVolume = this.debugShowBoundingVolume;
 
         var context = frameState.context;
 
@@ -634,6 +739,15 @@ define([
                 context : context
             });
             billboardCollection.textureAtlas = this._textureAtlas;
+        }
+
+        if (!defined(this._backgroundTextureAtlas)) {
+            this._backgroundTextureAtlas = new TextureAtlas({
+                context : context,
+                initialSize : whitePixelSize
+            });
+            backgroundBillboardCollection.textureAtlas = this._backgroundTextureAtlas;
+            addWhitePixelCanvas(this._backgroundTextureAtlas, this);
         }
 
         var uniformState = context.uniformState;
@@ -672,6 +786,7 @@ define([
         }
 
         this._labelsToUpdate.length = 0;
+        backgroundBillboardCollection.update(frameState);
         billboardCollection.update(frameState);
     };
 
@@ -711,6 +826,8 @@ define([
         this.removeAll();
         this._billboardCollection = this._billboardCollection.destroy();
         this._textureAtlas = this._textureAtlas && this._textureAtlas.destroy();
+        this._backgroundBillboardCollection = this._backgroundBillboardCollection.destroy();
+        this._backgroundTextureAtlas = this._backgroundTextureAtlas && this._backgroundTextureAtlas.destroy();
 
         return destroyObject(this);
     };